rename: Slide 폴더명 Carousel로 변경
