+++ frontend/src/pages/detail/components/Carousel/Carousel.tsx
