+++ frontend/src/pages/detail/components/Carousel/hooks/useSlide.ts
